Revert "Target CharSequence instead of String"
Targeting CharSequence is like targeting Object just becuase you can
call .toString on it get at the string.  This made the code quite a bit
uglier and the using of StringBuffer and StringBuilder is so rare that I
don't see the point in shielding the user from calling .toString herself.

This reverts commit 60b1db28971bc8a3a81fff1ba223692166bb4c17.

Conflicts:
	src/superstring/core.clj