Add translate
translate takes a translation map and translates all occurrences of the
keys in tmap into the corresponding values.  Optionally, it also takes a
set of characters to delete completely, prior to performing any mapping.

Stylistically the user might prefer mapping certain characters to nil,
instead of using the delete set, and this style is also supported